refactor: move reusable types from AuthContext to shared types file
- Move UserData, Message, Agent, WorkflowState, and Project interfaces to types.ts
- Update AuthContext imports to use shared types
- Keep AuthContextType in AuthContext as it's context-specific
